DjangoWithAI/requirements.txt
# FastAPI Gateway Requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
djangorestframework
djangorestframework-simplejwt
PyPDF2
google-generativeai
python-dotenv
requests
django-cors-headers
# Optional: For production deployment
# gunicorn==21.2.0
# redis==5.0.1  # For distributed rate limiting